Split binary out from AST
