+++ lib/CMakeLists.txt
@@ -1,5 +1,6 @@
 add_subdirectory(parser)
 add_subdirectory(ast)
+add_subdirectory(binary)
 add_subdirectory(codegen)
 
 separate_arguments(LLVM_DEFINITIONS_LIST NATIVE_COMMAND ${LLVM_DEFINITIONS})
