+++ lib/ast/CMakeLists.txt
@@ -3,9 +3,10 @@ set(CMAKE_CXX_FLAGS "${CMAKE_CXX_FLAGS} -g")
 
 add_library(AST
   AST.cpp
-  serializer.cpp
-  deserializer.cpp
 )
+
+target_link_libraries(AST
+  PUBLIC BinaryKore)
 
 install(
   TARGETS AST
