+++ lib/binary/CMakeLists.txt
@@ -1,0 +1,4 @@
+add_library(BinaryKore
+  serializer.cpp
+  deserializer.cpp
+)
