+++ lib/binary/deserializer.cpp
@@ -1,4 +1,4 @@
 #include <kllvm/ast/AST.h>
-#include <kllvm/ast/deserializer.h>
+#include <kllvm/binary/deserializer.h>
 
 namespace kllvm { } // namespace kllvm
