+++ tools/kore-convert/main.cpp
@@ -1,6 +1,6 @@
 #include "kllvm/ast/AST.h"
-#include "kllvm/ast/deserializer.h"
-#include "kllvm/ast/serializer.h"
+#include "kllvm/binary/deserializer.h"
+#include "kllvm/binary/serializer.h"
 #include "kllvm/parser/KOREParser.h"
 
 #include <llvm/Support/CommandLine.h>
